Add files via upload
This project contains comprehensive documentation for installing and configuring the FreeBSD operating system in a virtual environment using VirtualBox, along with a deep dive into the send() system call used in network programming. in addition to that it also included: history, file system support, and other core concepts of FreeBSD operating system.